refactor(Container): move styles
