+++ src/components/Container/Container.styles.ts
@@ -1,0 +1,12 @@
+import { WIDTHS } from '../../styles/variables'
+import styled from '@emotion/styled'
+
+const StyledContainer = styled.div`
+  position: relative;
+  margin-left: auto;
+  margin-right: auto;
+  width: auto;
+  max-width: ${WIDTHS.lg};
+`
+
+export default StyledContainer
